requirements.txt
loguru
pyyaml
requests
Pillow
transformers
pytesseract
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0